factored out serial and parallel williamson5 calculations in preparation for adding to tests
