+++ utils/shallow_water/__init__.py
@@ -1,3 +1,4 @@
 import utils.shallow_water.linear  # noqa: F401
 import utils.shallow_water.nonlinear  # noqa: F401
 import utils.shallow_water.williamson1992  # noqa: F401
+import utils.shallow_water.verifications  # noqa: F401
